fix(spotbugs): address multiple static analysis issues
- DLS_DEAD_LOCAL_STORE: Remove unused 'kind' variable in WitFunctionBinder
- DLS_DEAD_LOCAL_STORE: Remove unused 'currentTime' in MetadataCache
- RV_RETURN_VALUE_IGNORED: Use CharsetDecoder for UTF-8 validation in CustomSectionSecurity
- UPM_UNCALLED_PRIVATE_METHOD: Remove unused combineIssues method
- UMAC_UNCALLABLE_METHOD_OF_ANONYMOUS_CLASS: Remove inaccessible methods in FunctionContext
- ICAST_IDIV_CAST_TO_DOUBLE: Cast to double before division in GcStats
- ICAST_INTEGER_MULTIPLY_CAST_TO_LONG: Cast to long before multiply in FunctionStatistics
- RV_RETURN_VALUE_IGNORED_NO_SIDE_EFFECT: Use Math.sin result in ChaosEngineeringFramework

+++ wasmtime4j/src/main/java/ai/tegmentum/wasmtime4j/gc/GcStats.java
@@ -185,7 +185,7 @@ public final class GcStats {
     if (elapsed.isZero() || elapsed.isNegative()) {
       return 0.0;
     }
-    return bytesAllocated / elapsed.toNanos() * 1_000_000_000.0;
+    return (double) bytesAllocated / elapsed.toNanos() * 1_000_000_000.0;
   }
 
   /**
